server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.17
SQLAlchemy>=2.0.40
pymysql==1.1.1
cryptography==44.0.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
pydantic-settings==2.6.1
email-validator==2.2.0
